backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rpd-backend"
version = "0.1.0"
description = "Rpd Ganis GIU Backend - Emergent Cognitive Architecture"
requires-python = ">=3.10"
dependencies = [
    "langchain==0.1.0",
    "langchain-core==0.1.0",
    "langgraph==0.0.40",
    "langchain-anthropic==0.1.1",
    "langchain-google-genai==0.0.6",
    "pydantic-settings==2.1.0",
    "neo4j==5.15.0",
    "graphiti-core==0.3.0",
    "fastapi==0.108.0",
    "uvicorn==0.25.0",
]
